% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_two_grp.R
\name{ind_two_grp}
\alias{ind_two_grp}
\title{Independent Two Group T-Test}
\usage{
ind_two_grp(path, y1, y2)
}
\arguments{
\item{path}{The absolute path of file}

\item{y1}{One of the numeric variable names}

\item{y2}{The other numeric variable name}
}
\description{
This function allows you to apply independent two-group t-test on y1 and y2 where y1 and y2 are both numeric
}
\examples{
ind_two_grp(df, "numeric_col_1", "numeric_col_2")
ind_two_grp()
}
\keyword{t.test}

