% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_two_grp_w_factor.R
\name{ind_two_grp_w_factor}
\alias{ind_two_grp_w_factor}
\title{Independent Two Group T-Test With Factor}
\usage{
ind_two_grp_w_factor(path, y, x)
}
\arguments{
\item{path}{The absolute path of file}

\item{y}{The variable name of numeric value}

\item{x}{The variable name of binary factor value}
}
\description{
This function allows you to apply independent two-group t-test on y and x where y is numeric and x is a binary factor
}
\examples{
ind_two_grp_w_factor(df, "numeric_col", "binary_factor_col")
ind_two_grp_w_factor()
}
\keyword{t.test}

