% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample_t_test.R
\name{one_sample_t_test}
\alias{one_sample_t_test}
\title{One Sample T-Test}
\usage{
one_sample_t_test(path, y, mu)
}
\arguments{
\item{path}{The absolute path of file}

\item{y}{The numeric variable name}

\item{mu}{The mean of null hypothesis}
}
\description{
This function allows you to apply one sample t-test on y where y is numeric
}
\examples{
one_sample_t_test(df, "numeric_col")
one_sample_t_test()
}
\keyword{t.test}

