% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_t_test.R
\name{paired_t_test}
\alias{paired_t_test}
\title{Paired T-Test}
\usage{
paired_t_test(path, y1, y2)
}
\arguments{
\item{path}{The absolute path of file}

\item{y1}{One of the numeric variable names}

\item{y2}{The other numeric variable name}
}
\description{
This function allows you to apply paired t-test on y1 and y2 where y1 and y2 are both numeric
}
\examples{
paired_t_test(df, "numeric_col_1", "numeric_col_2")
paired_t_test()
}
\keyword{t.test}

